+++ test/__init__.py
@@ -5,7 +5,7 @@ import unittest
 from six import assertRegex, assertRaisesRegex
 
 from bfg9000.environment import Environment
-from bfg9000.path import abspath, Path, Root
+from bfg9000.path import abspath, Path, Root, InstallRoot
 
 __all__ = ['assertNotRegex', 'assertRaisesRegex', 'assertRegex', 'load_tests',
            'make_env', 'parameterize_tests', 'skip_if', 'skip_pred',
@@ -14,13 +14,14 @@ __all__ = ['assertNotRegex', 'assertRaisesRegex', 'assertRegex', 'load_tests',
 
 def make_env(platform=None, clear_variables=False, variables={}):
     args = (Path('bfgdir', Root.srcdir), None, None, abspath('srcdir'),
-            abspath('builddir'), {}, (False, False))
+            abspath('builddir'))
     if platform:
         with mock.patch('bfg9000.platforms.core.platform_name',
                         return_value=platform):
             env = Environment(*args)
     else:
         env = Environment(*args)
+    env.finalize({InstallRoot.prefix: abspath('prefix')}, (False, False))
 
     if clear_variables:
         env.variables = {}