+++ CHANGES.md
@@ -23,6 +23,8 @@
   input files when determining the builder
 - `args` and `kwargs` no longer supported for `build_step()`; use a lambda
   instead
+- Options specified on the command line (e.g. install locations) now override
+  settings in toolchain files
 
 ### Bug fixes
 - Include `options.bfg` in dist tarballs
@@ -31,6 +33,9 @@
   installed as symlinks
 - Generated header directories are now included as dependencies of compilation
   steps
+- Default install locations are now based on the *target* platform instead of
+  the host platform, with cross-platform builds defaulting to installation
+  disabled
 
 ---